refactor and add config options for future tunnel
