+++ config/poseidon.config
@@ -23,6 +23,8 @@ rabbit_port = 5672
 controller_log_file = /var/log/faucet/faucet.log
 controller_config_file = /etc/faucet/faucet.yaml
 controller_mirror_ports = {"switch1":3}
+tunnel_vlan = 999
+tunnel_name = poseidon_tunnel
 ignore_vlans = '[]'
 ignore_ports = '{}'
 trunk_ports = '{}'
@@ -59,5 +61,3 @@ bridge = opro0
 pipette_port = 6699
 pcap_location = /opt/poseidon/pcaps
 pcap_size = 50
-
-
